新增查询用户账单信息功能（完成99%）

+++ reckoning/src/main/java/cn/yugutou/reckoning/dto/resp/UserBillingInfoResp.java
@@ -3,14 +3,15 @@ package cn.yugutou.reckoning.dto.resp;
 import lombok.Data;
 import org.apache.ibatis.type.Alias;
 
+import java.math.BigDecimal;
 import java.util.Date;
 @Data
 @Alias("userBillingInfoResp")
 public class UserBillingInfoResp {
     //账单ID
-    private Integer billingId;
+    private Long billingId;
     //金额
-    private Integer amount;
+    private BigDecimal amount;
     //账单主题
     private String billTheme;
     //人数
